requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
elastic-transport==8.4.0
elasticsearch==8.5.3
Flask==2.2.2
fuzzywuzzy==0.18.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.2
pip==22.3.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
setuptools==65.5.0
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
selenium
kafka-python==2.0.2
pyspark==3.3.1